JS: Replaced Random Test Data
-Previously the random test data was a combination of random byte
vaalues from 0 to 255. While it made use of the whole convas, it was not
a linear set of values, so the end graph was hard to tell if it was
truly sorted.

Instead of random values, the data is now values 0-99 in a random order
to be sorted. This solves the problem as it will be sorted into a
perfect ascending line.