JS: isSorted no longer allows Equal Values
-Because the sorting functions are not loops and have been written to be
called interatively they do not have a condition that decides when they
stop in the function itself. Instead, they stop getting called once they
are detected to be sorted.

This caused a bug in the ShellStep function because the last two values
that need to be swapped are 99, and 98. It first sets 98 to 99, then the
next iteration it should set 99 to 98. But because when the iteration
exits it is left as 99, 99, isSorted detects this as sorted.

This function no longer accepts equal values so that shellSort will work
properly. This limits tests that are of random values.